+++ components/src/test/java/com/mercadolibre/android/andesui/badge/AndesBadgeIconPillTest.kt
@@ -3,6 +3,7 @@ package com.mercadolibre.android.andesui.badge
 import android.content.Context
 import android.os.Build
 import androidx.test.core.app.ApplicationProvider
+import com.mercadolibre.android.andesui.badge.hierarchy.AndesBadgeIconHierarchy
 import com.mercadolibre.android.andesui.badge.icontype.AndesBadgeIconType
 import com.mercadolibre.android.andesui.badge.size.AndesBadgePillSize
 import org.junit.Assert
@@ -27,13 +28,15 @@ class AndesBadgeIconPillTest {
 
     @Test
     fun `creating new badge with all parameters set by constructor`() {
+        val hierarchy = AndesBadgeIconHierarchy.LOUD
         val type = AndesBadgeIconType.SUCCESS
         val size = AndesBadgePillSize.LARGE
 
         andesBadgeIconPill = AndesBadgeIconPill(
             context,
             type,
-            size
+            size,
+            hierarchy
         )
 
         Assert.assertEquals(type, andesBadgeIconPill.type)
@@ -42,37 +45,44 @@ class AndesBadgeIconPillTest {
 
     @Test
     fun `creating new badge with parameters by default`() {
+        val defaultHierarchy = AndesBadgeIconHierarchy.LOUD
         val defaultType = AndesBadgeIconType.HIGHLIGHT
         val defaultSize = AndesBadgePillSize.SMALL
 
         andesBadgeIconPill = AndesBadgeIconPill(context)
 
+        Assert.assertEquals(defaultHierarchy, andesBadgeIconPill.hierarchy)
         Assert.assertEquals(defaultType, andesBadgeIconPill.type)
         Assert.assertEquals(defaultSize, andesBadgeIconPill.size)
     }
 
     @Test
     fun `creating new badge with parameters set by setter`() {
+        val hierarchy = AndesBadgeIconHierarchy.LOUD
         val type = AndesBadgeIconType.ERROR
         val size = AndesBadgePillSize.LARGE
 
         andesBadgeIconPill = AndesBadgeIconPill(context)
 
+        andesBadgeIconPill.hierarchy = hierarchy
         andesBadgeIconPill.type = type
         andesBadgeIconPill.size = size
 
+        Assert.assertEquals(hierarchy, andesBadgeIconPill.hierarchy)
         Assert.assertEquals(type, andesBadgeIconPill.type)
         Assert.assertEquals(size, andesBadgeIconPill.size)
     }
 
     @Test
     fun `creating new badge with default parameters from xml constructor`() {
+        val defaultHierarchy = AndesBadgeIconHierarchy.LOUD
         val defaultType = AndesBadgeIconType.HIGHLIGHT
         val defaultSize = AndesBadgePillSize.SMALL
         val attrs = Robolectric.buildAttributeSet().build()
 
         andesBadgeIconPill = AndesBadgeIconPill(context, attrs)
 
+        Assert.assertEquals(defaultHierarchy, andesBadgeIconPill.hierarchy)
         Assert.assertEquals(defaultType, andesBadgeIconPill.type)
         Assert.assertEquals(defaultSize, andesBadgeIconPill.size)
     }